src/test/java/org/apache/sling/feature/extension/apiregions/api/config/RangeTest.java
/*
 * Licensed to the Apache Software Foundation (ASF) under one or more
 * contributor license agreements.  See the NOTICE file distributed with
 * this work for additional information regarding copyright ownership.
 * The ASF licenses this file to You under the Apache License, Version 2.0
 * (the "License"); you may not use this file except in compliance with
 * the License.  You may obtain a copy of the License at
 *
 *      http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */
package org.apache.sling.feature.extension.apiregions.api.config;

import static org.junit.Assert.assertEquals;
import static org.junit.Assert.assertNull;
import static org.junit.Assert.assertTrue;

import java.io.IOException;

import javax.json.Json;

import org.apache.sling.feature.Extension;
import org.apache.sling.feature.ExtensionState;
import org.apache.sling.feature.ExtensionType;
import org.junit.Test;

public class RangeTest {

    @Test public void testClear() {
        final Range entity = new Range();
        entity.getAttributes().put("a", Json.createValue(5));
        entity.setMax(5);
        entity.setMin(20.1);
        entity.clear();
        assertTrue(entity.getAttributes().isEmpty());
        assertNull(entity.getMax());
        assertNull(entity.getMin());
    }

    @Test public void testFromJSONObject() throws IOException {
        final Extension ext = new Extension(ExtensionType.JSON, "a", ExtensionState.OPTIONAL);
        ext.setJSON("{ \"min\" : 5, \"max\" : 20.1 }");

        final Range entity = new Range();
        entity.fromJSONObject(ext.getJSONStructure().asJsonObject());
        assertEquals(5L, entity.getMin());
        assertEquals(20.1, entity.getMax());
    }

    @Test public void testToJSONObject() throws IOException {
        final Range entity = new Range();
        entity.setMin(5);
        entity.setMax(20.1);

        final Extension ext = new Extension(ExtensionType.JSON, "a", ExtensionState.OPTIONAL);
        ext.setJSON("{ \"min\" : 5, \"max\" : 20.1 }");

        assertEquals(ext.getJSONStructure().asJsonObject(), entity.toJSONObject());
    }
}